PythonFT/requirements.txt
numpy
Pillow
PyAutoGUI
datetime